Add "Replace queue" menu option to selection panels
If I want to listen to a whole album, I simply navigate to the album in
unplayer and touch one of the tracks. This replaces the play queue with
the chosen album. The same applies when choosing to listen to "All
tracks" of an artist. However, if I wish to listen to a few albums from
an artist, or a number of tracks from an album, I tend to navigate to
the artist or album in question, select the albums/tracks I want, and
then realise that there's no option to replace the play queue, so I have
to navigate back to the play queue, clear it, then navigate back to the
artist/album. This commit adds an option to replace the play queue with
the current selection in most of pages for selecting music.